+++ lib/streamlit/elements/map.py
@@ -23,7 +23,12 @@ from typing import TYPE_CHECKING, Any, Collection, Final, cast
 
 import streamlit.elements.deck_gl_json_chart as deck_gl_json_chart
 from streamlit import config, dataframe_util
-from streamlit.color_util import Color, IntColorTuple, is_color_like, to_int_color_tuple
+from streamlit.elements.lib.color_util import (
+    Color,
+    IntColorTuple,
+    is_color_like,
+    to_int_color_tuple,
+)
 from streamlit.errors import StreamlitAPIException
 from streamlit.proto.DeckGlJsonChart_pb2 import DeckGlJsonChart as DeckGlJsonChartProto
 from streamlit.runtime.metrics_util import gather_metrics
