+++ lib/streamlit/elements/vega_charts.py
@@ -60,9 +60,9 @@ from streamlit.util import HASHLIB_KWARGS
 if TYPE_CHECKING:
     import altair as alt
 
-    from streamlit.color_util import Color
     from streamlit.dataframe_util import Data
     from streamlit.delta_generator import DeltaGenerator
+    from streamlit.elements.lib.color_util import Color
 
 # See https://vega.github.io/vega-lite/docs/encoding.html
 _CHANNELS: Final = {
